Added joystick routine at 2100
Replaced INPUT Z$ with joystick routine that assigns Z$ according to the direction the stick is moved.  This provides N, S, E, W results.

When the joystick button is pressed a different set of values are returned.  F, Q for Up/Down.  And Y, N (Yes/No) for Left/Right.